requirements.txt
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
Django==5.2.5
dnspython==1.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pymongo==4.14.0
python-dotenv==1.1.1
python-telegram-bot==22.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
